requirements.txt
shodan
cmd2
python-nmap
keyring